+++ R/backports.R
@@ -1,7 +1,7 @@
 
 if (getRversion() < "3.2.0") {
   dir.exists <- function(paths) {
-    info <- file.info(paths, extra_cols = FALSE)
+    info <- renv_file_info(paths)
     info$isdir %in% TRUE
   }
 }